refactor: 외부 서비스 기반의 로직 흐름 중심 3-layer 구조로 변경
기존 구조는 API 중심의 라우팅 흐름(Route → Service → Provider)으로 개발했으나,
현재 프로젝트는 Discord, Firebase 등 외부 서비스 중심의 비동기 흐름이 주가 되어 구조 상 제약이 있어 구조를 변경

이에 따라 Route 계층 대신 Application 계층을 도입하여 주요 로직의 진입점과 흐름을 담당하도록 변경,
Service 계층은 비즈니스 로직을 처리,
Provider는 외부 서비스와의 통신만을 전담하도록 역할을 명확히 분리
