--- functions/src/services/discordService.ts
@@ -1,3 +1,0 @@
-import { ResponseRecruitData } from "../types/responseRecruitData";
-
-export const discordService = async function(list: ResponseRecruitData[]) {};
