+++ functions/src/types/global.d.ts
